library_project/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.22,<1.0
django-simple-history>=3.1,<4.0